refactor: Separated concerns
